refactor: move random to utils
