+++ libs/base/utils/src/js/noop.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/no-empty-function
+export const NOOP = () => {};
