+++ libs/base/utils/tsconfig.json
@@ -1,4 +1,7 @@
 {
   "extends": "../../../tsconfig.base.json",
-  "include": ["src/**/*.ts"]
+  "include": ["src/**/*.ts"],
+  "compilerOptions": {
+    "types": ["vitest/globals"]
+  }
 }
