+++ libs/core/app/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./questions/get-new-variant";
 export * from "./questions/get-question";
 export * from "./questions/get-question-preview";
-export * from "./random/random";
 export * from "./tests/create-test";
 export * from "./tests/process-test";
 export * from "./tests/get-number-of-available-questions";
