+++ libs/core/github/src/functions/create-new-question-pr.ts
@@ -1,7 +1,7 @@
 import * as babelPlugin from "prettier/plugins/babel";
 import * as estreePlugin from "prettier/plugins/estree";
 import { format } from "prettier/standalone";
-import { getRandomId } from "@chair-flight/core/app";
+import { getRandomId } from "@chair-flight/base/utils";
 import { getOctokit } from "../config/oktokit";
 import type { QuestionBankQuestionTemplate } from "@chair-flight/base/types";
 import type { questionEditSchema } from "@chair-flight/core/schemas";
