+++ libs/trpc/server/src/routers/containers/flashcards-router.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { getRandomShuffler } from "@chair-flight/core/app";
+import { getRandomShuffler } from "@chair-flight/base/utils";
 import { questionBanks } from "@chair-flight/core/question-bank";
 import { questionBankNameSchema } from "@chair-flight/core/schemas";
 import { publicProcedure, router } from "../../config/trpc";
